medição de tempo

+++ src/file.h
@@ -2,12 +2,12 @@
 #define TP_02_F_H
 #endif
 
-long int **readFileIntoMatrix(char *filepath, long int *rows, long int *cols);
+long long int **readFileIntoMatrix(char *filepath, long long int *rows, long long int *cols);
 
-char *generateRandomFile(long int *rows, long int *cols);
+char *generateRandomFile(long long int *rows, long long int *cols);
 
-long int **initializeMatrix(long int rows, long int cols);
+long long int **initializeMatrix(long long int rows, long long int cols);
 
-char *generateRandomFileWithInput(long int *rows, long int *cols, long int *range);
+char *generateRandomFileWithInput(long long int *rows, long long int *cols, long long int *range);
 
 void flushIn();
